Collapsed Mac/Win layers for dactyl_manuform_5x6_5_davidf
Now we don't have separate layers for Mac/Win modes, because so few keys differ between them.
Instead we default to Win and switch to Mac mode using a user key. Mac layout changes are hard-coded because there are so few of them and they never change. This method is easier to maintain and more memory efficient.